GTC-1865 Pagination Datasets (#327)
* :sparkles: feat(paginate): Create first tests and params to `get_datasets`

- `get_datasets` uses raw sql. I played a bit with converting it to a real sqlalchemy call and failed to have a working translation. I went ahead and started to do _the simplest possible thing_ to get started returning paginated rows.
- Started with a test suite to flesh out an implementation and work more with `unittest.mock` and `pytest`


feat(paginate): Create first couple tests and default params to `get_datasets`

- `get_datasets` uses raw sql. I played a bit with converting it to a real sqlalchemy call and failed to have a working translation. I went ahead and started to do _the simplest possible thing_ to get started returning paginated rows.
- Started with a test suite to flesh out an implementation and work more with `unittest.mock` and `pytest`


* :sparkles: feat(paginate): Use `paginate_datasets` in the `datasets` controller

- I'm still adapting my understanding of pythonic `async/await` and what is required to support it in tests.
- Pay special attention to the use of the words `spy` and `stub` when dealing with collaborators.
  - `spy` is used to verify that I'm calling the collaborator correctly. The test cares about nothing else. The test will fail only if the function under tests no longer upholds its part of the contract
  - `stub` is used to verify that the function under test behaves property when provided the agreed upon value of the collaborating function. It will fail only if the function under test fails to uphold its contract with callers. The `stub` means that the collaborator behavior isn't under scrutiny. Only that it continues to return an agreed upon value for the scenario at hand.


* :sparkles: feat(paginate): Introduce `meta` section for paginated results

I don't care, at all, what my collaborator is doing for this first set of `meta` tests. Therefore, they are declared as `dummy` test doubles.


* :sparkles: feat(paginate): Add `size`, `total_items`, and `total_pages` to meta

- Definitely not a final implementation here but just enough to get started. I do not like having the two crud functions in the `paginate_datasets` formal parameters.


* :sparkles: feat(paginate): Add paginated response to datasets controller

- This is just the first cut. I have to learn about formatting json properties from pydantic models.


* :sparkles: feat(paginate): Calculate the correct offset when paging

Needed to convert the concept of a page to an actual offset for use with PostgreSQL queries.


* :art: refactor(paginate): Group tests into their own files


* :sparkles: feat(paginate): Introduce pagination links


* :sparkles: feat(paginate): Add link refs

Links property of a paginated request now returns:
- self
- first
- last
- prev
- next


* :art: refactor(paginate): Use `typing.NamedTuple` for links


* :art: refactor(paginate): Use `typing.NamedTuple` for meta


* :sparkles: feat(paginate): Route unpacks `data`, `link`, and `meta`


* :sparkles: feat(paginate): Add boundary conditions for links


* :art: refactor(paginate): Rename function params and clean up logic


* :art: refactor(paginate): Move `ValueError` exceptions up

Also, added a `Meta` section test for when there are no records returned. We still want `total_pages` to be `1`


* :art: refactor(paginate): Flip the condition for choosing a response

I went back and made the predicate more positive sounding instead of mucking it up with `not`s


* :white_check_mark: test(paginate): Add an end-to-end test for `/datasets`

This is just getting up and running with the end-to-end tests. More to follow. :smile:


* :white_check_mark: test(paginate): Test for pagination query params

This is just getting up and running with the end-to-end tests. More to follow. :smile:


* :sparkles: feat(paginate): Add query params to `/datasets` endpoint

The `datasets` controller now calls `paginate` with arguments from the query parameters.


* :sparkles: feat(paginate): Ensure that query is ordered for consistent paging

From PostgreSQL reference:
"When using LIMIT, it is important to use an ORDER BY clause that constrains the result rows into a unique order. Otherwise you will get an unpredictable subset of the query's rows. You might be asking for the tenth through twentieth rows, but tenth through twentieth in what ordering? The ordering is unknown, unless you specified ORDER BY.

 The query optimizer takes LIMIT into account when generating query plans, so you are very likely to get different plans (yielding different row orders) depending on what you
 give for LIMIT and OFFSET. Thus, using different LIMIT/OFFSET values to select different subsets of a query result will give inconsistent results unless you enforce a
 predictable result ordering with ORDER BY. This is not a bug; it is an inherent consequence of the fact that SQL does not promise to deliver the results of a query in any
 particular order unless ORDER BY is used to constrain the order."

https://www.postgresql.org/docs/current/queries-limit.html


* :sparkles: feat(paginate): Configure validation for page query params


* :sparkles: feat(paginate): Page number that is too large returns 4xx

Page number is a value passed in by the client. We can't know if the page number is too large until we get the datasets count. If it's past the max pages then we let the client know about it so they can try to fix it.


* :sparkles: feat(paginate): Default the page size to 10

https://resource-watch.github.io/doc-api/concepts.html#pagination


* :memo: docs(paginate): Add API docs


* :art: refactor(paginate): Make pagination function generic

I wanted to remove the dataset dependency so it can be used for all of our collections.


* :art: refactor(paginate): Move paginate.py to `app/utils`

Seems like a much better home for this functionality.


* :art: refactor(paginate): Make imports absolute


* :art: refactor(paginate): Clean up test code

Remove unneeded test fixture usage and in-line `return_value` in some Mocks for consistency.


* :art: refactor(paginate): Use Pydantic instead of NamedTuple

This brings the pagination objects in-line with the rest of the codebase.

* :art: refactor(paginate): Rename `paginate` to `pagination`


* :art: refactor(paginate): Add missing return types in `paginate` module


* :art: refactor(paginate): Remove sketchy pagination link logic

The 'if' conditions at the end of a statement are hard to read.
Also, DRY up the repeated `f` template for link refs.


* :memo: docs(paginate): Update the `/datasets` endpoint description


* :art: refactor(paginate): Do not always attempt pagination at `datasets` endpoint

It is much cleaner to only try to paginate when pagination parameters are passed via query params.


* :art: refactor(paginate): `paginate.py` is only used when pagination is requested

Previously, `paginate.py` tried to hide the fact that it would not paginate if you didn't pass `size` or `page`. Unfortunately, this gave the function two big jobs. Letting it focus on just doing one thing actually simplified things here and in the path operation function `/datasets`.


* :memo: docs(paginate): Realign the `Pagination Meta` titles in `Responses.py`

`total_items` and `total_pages` got swapped somehow.


* :art: refactor(paginate): Remove redundant `Links` and `Meta` pydantic types

Originally, I had just some named tuples for links and meta pagination information. Once we decided to move away from that, it got overlooked. Also, I held on the thought that I might need a mapping from the data and the pydantic response. It was never needed - YAGNI!