GTC-1831 Assets Pagination (#329)
* :white_check_mark: test(paginate-assets): Add initial test for `AssetsResponse`

* :sparkles: feat(paginate-assets): Add `PaginatedAssetsResponse`

* :art: refactor(paginate-assets): Separate query building from execution
This allows for the use of the query in the necessary pagination functions.

* :sparkles: feat(paginate-assets): Introduce pagination functions
These are higher order functions so that we can honor the interface required in `paginate.py`

* :sparkles: feat(paginate-assets): Add pagination functionality to `/assets`

* :sparkles: feat(paginate-assets): Serialize the ORMAsset to a Pydantic entity
We do this for creating the proper Metadata objects for the response payload.

* :memo: doc(paginate-assets): Add pagination behavior to route description

* :memo: doc(paginate-assets): Make `PaginationLinks` examples more generic

* :memo: docs(paginate-asset): Add docs to `Assets` crud functions

* :art: refactor(paginate-assets): Rearrange function order in `crud/assets`

* :art: refactor(paginate-assets): Remove usage of `get_all_assets`

The same functionality can be achieved with the `get_assets_by_filter`. Additionally,
`get_all_assets` was only being used to support test cases.

* :fire: refactor(paginate-assets): Remove `get_all_assets` from `crud/assets`

* :art: refactor(paginate-assets): Remove usage of `get_assets_by_type`
The same functionality can be achieved with the `get_assets_by_filter`. Additionally,
`get_assets_by_type` was only being used to support test cases.

* :fire: refactor(paginate-assets): Remove `get_assets_by_type` from `crud/assets`