GTC-2443 Fix /fields query for version with only rasters
Uncommented and fixed some existing code. Added in the extra magic field
"area__ha", as requested.

Added a test. I wasn't able to patch an asset of the
generic_raster_version to a values_table with some rows, so I wasn't
able to test the values_table value. To make the test work, I made sure
the generic_raster_version was marked as the latest version.