GTC-2161 Document special fields for raster-only datasets
Import the documentation for query of raster-only datasets. In
particular, mention that the special fields 'area__ha', 'latitude', and
'longitude' can be used.

Also, now that we know that 'latitude' and 'longitude' can be used for
all raster-only datasets, add them to the actual /fields query (along
with 'area__ha', which I added in a previous commit).