GTC-2649 Include asset_id in the version get endpoint
The version get endpoint already includes asset_type and asset_uri, but
it would be more useful if it also included the associated asset_id.

I believe it is fine to modify Version. In some cases (like
add_new_version) where a new asset is being created in the background,
the _version_response may have an asset with an empty asset_id, but this
seems to be all fine (all tests pass). But let me know if you think I
should leave Version unmodified and create a separate type only for the
output of get_inversion that includes the asset_id.