GTC-2708 Expire all tile cache files when deleting a dataset version
We currently delete all datalake files, but have not been deleting the
tile cache files in s3://gfw-tiles. The expire_s3_objects() function
seems to work, since it is already used when deleting only a static
vector tile cache or a raster tile cache.

Now that expire_s3_objects is being called when we delete a version, I
needed to add in a monkeypatch for expire_s3_objects() in
create_vector_source_version() and generic_raster_version() for the
tests_v2 tests.