GTC-2822 Check for dataset ownership on asset requests
Make use of get_owner in the asset operations.

The checks for the pure asset operations were slightly trickier than
dataset/version operations, because the dataset name is available in the
URL path. You need to do a DB operation to get the dataset name. So, I
put the actual get_owner() call in the main function rather than in a
Depends() argument. I didn't try to add a new Depends() function that
magically does the DB operation, since some of the functions already do
the DB operation for other reasons.

Updated the test for update_metadata and update_field_metadata to test
success and failure cases for different users.

Added a missing check for the ADMIN role in get_owner() and updated its
comment.

Updated the detail of the various permission exceptions to be more
informative (GTC-2795). Let me know what you think. We could put a bit
more information in these messages if we did these errors in the main
functions, rather than in the Depends functions, but that is probably
not worth it.

In test_dataset.py, several tests were being skipped unintentionally
because they were missing the '@pytyest.mark.asyncio' annotation, so I
added that in where needed.