GTC-2788 Do proper serialization for UUIDs
This is in case there is a unusual query that does SELECT(*) on a
dataset that has a UUID field. Currently, we get a 500 with an error
"TypeError: Type is not JSON serializable: asyncpg.pgproto.pgproto.UUID".

We need to force UUID to str so it can be serialized, just as we are
already doing with Decimal.

https://stackoverflow.com/questions/64395136/typeerror-object-of-type-uuid-is-not-json-serializable