GTC-2864 Make test output for Data API be less verbose
I change both script/test and script/test_v2 to not show warnings or
do/show code coverage by default. So, the actual test results won't
scroll off the screen in the output when running a single or few tests.
Also, I added an option to cause stdout of the tests to be displayed, so
that you can see debugging print() calls when running tests, even if the
test succeeds. (Normally all output of a test is captured and not
displayed if the test succeeds.)

New options (all off by default):
  --do-cov - Run code coverage and display it after the test results
  --show-warning - Show python warnings when compiling tests
  --show-stdout - Show output of successful tests