Change specification of tests to include tests/ or tests_v2/ path
If you are specifying specific test files for scripts/test or
scripts/test_v2, now include the relative catch from the the top of the
repo (don't drop tests/ or tests_v2). Will make file completion of the
test file names from the shell be usable.

Also, enable --show-warnings for the the tests run in the Github
actions.