GTC-2958 Tag Docker image from docker hash, so it always exists
Currently, we tag the new Docker image with the current Git SHA. But the
container_registry module only creates a new Docker image if the docker
contents change. So, if the docker contents haven't changed with this
Git change, we can have a bug where we reference the app docker via a
tag (of the new Git SHA) which doesn't exist.

The fix is to use instead as a tag the hash of the docker contents. We
use the same hash script that the container_registry module uses.
Therefore, we will always be using a tag that exists, either because the
container_register module just pushed a new docker with the new tag, or
the docker already exists under the docker hash tag, because the docker
contents and has haven't changed.

I noticed a bunch of bugs in the container_registry module's hash.sh
script, which I will fix later. One of the main things is that it
doesn't ignore comments, so it can match on words in the comments. For
that reason, I removed the '# Docker Files' comment, which was causing
the Dockerfile itself to be ignored during hashing.