Sort the assets by creation time in the version endpoint, just like /assets
We made the asset list in the GET version endpoint more useful recently
by including the asset-id. But that list is not being sorted by creation
time, unlike the list provided by /assets, so the assets are somewhat
randomly ordered.

So, add the change to sort the asset list of the GET version endpoint by
creation time. This ensures that the default asset is always first, etc.