GTC-2708 Fix deleting of the tile cache when deleting a version
The main issue was that the correct lifecycle action permission is
actually s3:PutLifecycleConfiguration, but we've always had
s3:PutBucketLifecycleConfiguration in the policy templates.

A second issue is that the expiration time for a bucket/folder must be a
date specified as midnight UTC. We were just using the current time, so
that caused an error when trying to set the lifecycle rule.

In both cases, we weren't seeing the error, because delete_all_assets
was called from a background task, so any exception was getting lost
without being logged. So, I added a try-except around the call to
put_bucket_lifecycle_configuration(), so we log future errors (and don't
kill the task).

Given these fixes, I was able to re-enable the use of
expire_s3_objects() and it worked!