GTC-3276 Change calculation of extent metadata using wgs84Extent
The existing calculation was sometimes changing longitude 180 to -180
and vice versa. 180 and -180 are not equivalent, since a bounding box
always takes the shortest direction around, which is different for 180
(or 179.9999) and -180 (or -179.9999).

Change the calculation using to_4326() on cornerCoordinates (which is in
meters) to existing wgs84Extent, which is already exactly in degrees.
The resulting feature coordinates match the values from before the
switch from using pixetl.