GTC-3375 Add "copy_solo_tiles" option to optimize creation of int-dist alerts raster (#710)
* GTC-3375 Add "copy_solo_tiles" option to optimize creation of int-dist alerts

When we merge integrated alerts and dist alerts rasters, much of the
globe has only dist alerts. So, we want to optimize the process by
copying the dist alerts tiles directly when to the final raster if there
is no corresponding integrated alerts tile. We provide the
copy_solo_tiles option to do this, which runs the "copy_solo_tiles.sh"
script after the main raster is created with "union_bands = False". The
script also correctly updates extent.geojson and tiles.geojson.

* Validate that auxiliary_asset is not provided if source_uri is provided.