Try to get tags on batch jobs propagated to the Cost Explorer (#714)
Tim suggested we were missing setting propagate_tags in our
aws_batch_job_definition resource in the batch module, and also for Data
API, we never set propagateTags for the batch job itself. We do
correctly propagate tags for EMR jobs.

Hopefully, this will allow tags such as the dataset to be propagated to
the Cost Explorer. And we could also add tags for the job type (such as
COGify vs create_raster, etc.)